+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Implement the `drag` and `zoom` plugins
+
 v0.2.0
 ------